requirements.txt
Flask==2.2.2
pytest==7.2.0
report_monaco_road==0.0.9
Flask-RESTful==0.3.9
dicttoxml~=1.7.4
flasgger~=0.9.5
flake8 ==5.0.4
isort == 5.10.1
pre-commit == 2.20.0
